refactor mapToPrismaOrderBy
